requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-baton==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-model-utils==4.2.0
django-ninja==0.18.0
Faker==13.13.0
gunicorn==20.1.0
mimesis==5.4.0
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
